pyproject.toml
[project]
name = "aptify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "django-allauth[socialaccount]>=65.14.0",
    "django-ninja>=1.5.3",
    "whitenoise>=6.11.0",
    "psycopg[binary]>=3.2",
    "dj-database-url>=1.0.0",
    "django-environ>=0.10.0",
    "django-environ>=0.12.0",
    "pillow>=12.1.0",
    "django-allauth-ui>=1.8.1",
    "django-widget-tweaks>=1.5.1",
    "slippers>=0.6.2",
    "gunicorn>=24.1.1",
]
